workflows-python/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-cosmos
azure-functions
azure-storage-blob
azure-cosmosdb-table
numpy
Pillow